#patch refactor types constants utils
