--- src/constants/enums/index.ts
@@ -1,1 +1,0 @@
-export { EAlign } from "./alignment";
