--- src/constants/types/common.ts
@@ -1,36 +1,0 @@
-// Content Direction
-export type ContentDirection = "ltr" | "rtl" | "auto";
-// Container Layout
-export type ContainerLayout = "fluid" | "swipe" | "default";
-// Display Zone
-export type DisplayZone =
-	| "top"
-	| "right"
-	| "bottom"
-	| "left"
-	| "top-right"
-	| "bottom-right"
-	| "bottom-left"
-	| "top-left"
-	| "center";
-
-// Heading Size
-export type HeadingSize = "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
-
-// Grid TypeType
-export type GridType = "default" | "masonry";
-
-// Media type
-export type MediaType = "image" | "video" | "audio";
-
-// Message types
-export type MessageType = "info" | "warning" | "error";
-
-// Content Layout
-export type ObjectLayout = "horizontal" | "vertical" | "variant";
-
-// Profile hide
-export type ProfileHidden = "name" | "nickName" | "avatar" | "info";
-
-// Ratio
-export type Ratio = "1x1" | "1x2" | "3x2" | "3x4" | "4x3" | "9x16" | "16x9" | "16x10";
