--- src/constants/types/list.ts
@@ -1,4 +1,0 @@
-// List type
-export type List = "unordered" | "ordered" | "description";
-export type UnorderedListMarker = "none" | "disc" | "circle" | "square";
-export type OrderedListType = "1" | "A" | "a" | "I" | "i";
