--- src/constants/types/size.ts
@@ -1,4 +1,0 @@
-// Size
-export type Size = "tiny" | "small" | "medium" | "large" | "large-x" | "large-xx" | "large-xxx";
-// Size plus
-export type SizePlus = "xs" | "sm" | "md" | "lg" | "xl" | "xxl" | "xxxl" | "o-xxxl" | "xo-xxxl";
