+++ src/index.css
@@ -1,5 +1,4 @@
 :root {
-  --overflow-hidden: hidden;
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
