+++ src/utils/CheckType.ts
