--- src/utils/function/index.ts
@@ -1,1 +1,0 @@
-export { isObject, isStringOrStringArray } from "./CheckType";
